modulized button component styles
